chore(types): codegen with useEnhancedDecimal
Not really "enhanced". See https://github.com/hyperweb-io/telescope/issues/652
