+++ packages/cosmic-proto/src/codegen/agoric/vbank/vbank.ts
@@ -1,7 +1,8 @@
 //@ts-nocheck
 import { Coin, type CoinSDKType } from '../../cosmos/base/v1beta1/coin.js';
 import { BinaryReader, BinaryWriter } from '../../binary.js';
-import { Decimal, isSet } from '../../helpers.js';
+import { Decimal } from '../../decimals.js';
+import { isSet } from '../../helpers.js';
 import { type JsonSafe } from '../../json-safe.js';
 /** The module governance/configuration parameters. */
 export interface Params {
