+++ packages/cosmic-proto/src/codegen/cosmos/base/v1beta1/coin.ts
@@ -2,6 +2,7 @@
 import { BinaryReader, BinaryWriter } from '../../../binary.js';
 import { isSet } from '../../../helpers.js';
 import { type JsonSafe } from '../../../json-safe.js';
+import { Decimal } from '../../../decimals.js';
 /**
  * Coin defines a token with a denomination and an amount.
  *
@@ -162,7 +163,9 @@ export const DecCoin = {
       writer.uint32(10).string(message.denom);
     }
     if (message.amount !== '') {
-      writer.uint32(18).string(message.amount);
+      writer
+        .uint32(18)
+        .string(Decimal.fromUserInput(message.amount, 18).atomics);
     }
     return writer;
   },
@@ -178,7 +181,7 @@ export const DecCoin = {
           message.denom = reader.string();
           break;
         case 2:
-          message.amount = reader.string();
+          message.amount = Decimal.fromAtomics(reader.string(), 18).toString();
           break;
         default:
           reader.skipType(tag & 7);
@@ -292,7 +295,7 @@ export const DecProto = {
     writer: BinaryWriter = BinaryWriter.create(),
   ): BinaryWriter {
     if (message.dec !== '') {
-      writer.uint32(10).string(message.dec);
+      writer.uint32(10).string(Decimal.fromUserInput(message.dec, 18).atomics);
     }
     return writer;
   },
@@ -305,7 +308,7 @@ export const DecProto = {
       const tag = reader.uint32();
       switch (tag >>> 3) {
         case 1:
-          message.dec = reader.string();
+          message.dec = Decimal.fromAtomics(reader.string(), 18).toString();
           break;
         default:
           reader.skipType(tag & 7);
