+++ packages/cosmic-proto/src/codegen/cosmos/staking/v1beta1/staking.ts
@@ -11,13 +11,9 @@ import {
 } from '../../../google/protobuf/duration.js';
 import { Coin, type CoinSDKType } from '../../base/v1beta1/coin.js';
 import { BinaryReader, BinaryWriter } from '../../../binary.js';
-import {
-  isSet,
-  Decimal,
-  fromJsonTimestamp,
-  fromTimestamp,
-} from '../../../helpers.js';
+import { isSet, fromJsonTimestamp, fromTimestamp } from '../../../helpers.js';
 import { type JsonSafe } from '../../../json-safe.js';
+import { Decimal } from '../../../decimals.js';
 /** BondStatus is the status of a validator. */
 export enum BondStatus {
   /** BOND_STATUS_UNSPECIFIED - UNSPECIFIED defines an invalid validator status. */
