+++ packages/cosmic-proto/src/codegen/cosmos/staking/v1beta1/tx.ts
@@ -12,13 +12,9 @@ import {
   type TimestampSDKType,
 } from '../../../google/protobuf/timestamp.js';
 import { BinaryReader, BinaryWriter } from '../../../binary.js';
-import {
-  isSet,
-  Decimal,
-  fromJsonTimestamp,
-  fromTimestamp,
-} from '../../../helpers.js';
+import { isSet, fromJsonTimestamp, fromTimestamp } from '../../../helpers.js';
 import { type JsonSafe } from '../../../json-safe.js';
+import { Decimal } from '../../../decimals.js';
 /** MsgCreateValidator defines a SDK message for creating a new validator. */
 export interface MsgCreateValidator {
   description: Description;
