+++ packages/cosmic-proto/src/codegen/index.ts
@@ -1,6 +1,6 @@
 //@ts-nocheck
 /**
- * This file and any referenced files were automatically generated by @cosmology/telescope@1.8.3
+ * This file and any referenced files were automatically generated by @cosmology/telescope@1.12.8
  * DO NOT MODIFY BY HAND. Instead, download the latest proto files for your chain
  * and run the transpile command or npm scripts command that is used to regenerate this bundle.
  */
@@ -17,6 +17,7 @@ export * from './icq/bundle.js';
 export * from './ics23/bundle.js';
 export * from './stride/bundle.js';
 export * from './tendermint/bundle.js';
+export * from './decimals.js';
 export * from './json-safe.js';
 export * from './varint.js';
 export * from './utf8.js';
