+++ packages/cosmic-proto/src/codegen/varint.ts
@@ -1,6 +1,6 @@
 //@ts-nocheck
 /**
- * This file and any referenced files were automatically generated by @cosmology/telescope@1.8.3
+ * This file and any referenced files were automatically generated by @cosmology/telescope@1.12.8
  * DO NOT MODIFY BY HAND. Instead, download the latest proto files for your chain
  * and run the transpile command or npm scripts command that is used to regenerate this bundle.
  */
